src/sharepoint/client.py
from typing import List, Dict, Any, Optional
from azure.identity import ClientSecretCredential
from msgraph import GraphServiceClient
from msgraph.generated.sites.item.lists.item.items.items_request_builder import ItemsRequestBuilder
from kiota_abstractions.base_request_configuration import RequestConfiguration

from ..config import settings

class SharePointClient:
    def __init__(self):
        self._credential = None
        self._client = None

    @property
    def credential(self):
        if self._credential is None:
            self._credential = ClientSecretCredential(
                tenant_id=settings.TENANT_ID,
                client_id=settings.CLIENT_ID,
                client_secret=settings.CLIENT_SECRET,
            )
        return self._credential

    @property
    def client(self):
        if self._client is None:
            self._client = GraphServiceClient(
                self.credential, scopes=["https://graph.microsoft.com/.default"]
            )
        return self._client

    async def get_site_lists(self, site_id: str) -> List[Dict[str, Any]]:
        """Get all lists in a site."""
        # This is a simplified wrapper. In a real implementation, we'd use the generated models.
        # But for quick prototyping, raw request execution or using the SDK models is fine.
        # Using SDK models is preferred but verbose.
        
        # Getting lists
        lists = await self.client.sites.by_site_id(site_id).lists.get()
        return lists.value if lists and lists.value else []

    async def get_list_columns(self, site_id: str, list_id: str) -> List[Dict[str, Any]]:
        """Get columns for a specific list."""
        columns = await self.client.sites.by_site_id(site_id).lists.by_list_id(list_id).columns.get()
        return columns.value if columns and columns.value else []

    async def get_list_item_count(self, site_id: str, list_id: str) -> int:
        """Get the number of items in a list."""
        request_config = ItemsRequestBuilder.ItemsRequestBuilderGetRequestConfiguration(
            query_parameters=ItemsRequestBuilder.ItemsRequestBuilderGetQueryParameters(
                top=1,
            )
        )
        result = await self.client.sites.by_site_id(site_id).lists.by_list_id(list_id).items.get(request_configuration=request_config)
        if result and result.odata_count is not None:
            return result.odata_count
        # Fallback: fetch all items and count (expensive but works)
        items = await self.get_list_items(site_id, list_id)
        return len(items)

    async def get_list_items(self, site_id: str, list_id: str, top: int = 200, expand: str = "fields") -> List[Dict[str, Any]]:
        """Get items from a list with pagination."""
        request_config = ItemsRequestBuilder.ItemsRequestBuilderGetRequestConfiguration(
            query_parameters=ItemsRequestBuilder.ItemsRequestBuilderGetQueryParameters(
                top=top,
                expand=[expand]
            )
        )
        
        items = await self.client.sites.by_site_id(site_id).lists.by_list_id(list_id).items.get(request_configuration=request_config)
        
        all_items = []
        if items and items.value:
            all_items.extend(items.value)
            
            # Handle pagination
            next_link = items.odata_next_link
            while next_link:
                # Create a new request using the next link
                # The SDK models don't make this super easy to just "get next", 
                # often we need to parse or use a raw request.
                # Here we use the with_url method if available on request builder, 
                # or just raw execute.
                
                # Simplified: limit to 5000 items to prevent infinite loops during dev
                if len(all_items) >= 5000:
                    break
                    
                # Creating a request for next page
                # This requires constructing a request builder from the URL
                # For this MVP, let's assume single page or small lists.
                # Proper implementation requires:
                # next_page = await self.client.sites...with_url(next_link).get()
                # But strict SDK typing makes dynamic builder hard.
                # Let's use the raw client capability if possible or skip for now.
                break
                
        return all_items

    async def get_site_id_by_url(self, site_url: str) -> str:
        """Resolve site URL to site ID (hostname,sppath,guid)."""
        # Logic to parse URL and query graph
        # For now, assume site_id is passed directly or implement hostname lookup
        # standard format: hostname:/sites/sitename
        from urllib.parse import urlparse
        parsed = urlparse(site_url)
        hostname = parsed.hostname
        path = parsed.path.strip("/")
        
        site = await self.client.sites.by_site_id(f"{hostname}:/{path}").get()
        return site.id

# Singleton instance
sharepoint_client = SharePointClient()
